Make icomplete-vertical-separator customisable
This can allow users to adapt the separator to their particular needs.
Such as:

    (setq icomplete-vertical-separator "\n······\n")
    (setq icomplete-vertical-separator "\n——————\n")

One could want that to better delineate candidates that are naturally
long or multi-line: e.g. `kill-ring'.

It would, thus, allow the user to maintain "wrapper" functions that set
this variable to the appropriate value, depending on the specifics of
the command.  For instance, to use "\n" in `M-x' and one of the
aforementioned values in a `kill-ring' item selection.